requirements.txt
# API and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Trading and Analysis
ccxt==4.1.13
pandas==2.1.3
numpy==1.26.2
ta==0.10.2
scipy==1.11.4

# Content Generation
openai==1.3.5
langchain==0.0.350
python-docx==1.0.1
markdown==3.5.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Monitoring and Logging
sentry-sdk==1.32.0
structlog==23.2.0
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
aiohttp==3.9.1
tenacity==8.2.3 